requirements.txt
# Core ML and Computer Vision libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Inference and Model Serving
inference-gpu>=0.32.0
onnxruntime-gpu>=1.18.0

# Sports Analysis (Roboflow)
# Note: Install with: pip install git+https://github.com/roboflow/sports.git
# sports @ git+https://github.com/roboflow/sports.git

# Supervision library for computer vision
supervision>=0.18.0

# Progress bars and utilities
tqdm>=4.65.0

# Jupyter and interactive development (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Video processing and visualization
matplotlib>=3.7.0
pillow>=10.0.0

# Additional utilities
filterpy>=1.4.5  # For tracking algorithms (if needed)
scipy>=1.10.0  # For spatial calculations (ConvexHull, etc.) 